+++ DataFormats/L1TMuon/src/L1MuBMSecProcId.cc
@@ -15,7 +15,7 @@
 // This Class's Header --
 //-----------------------
 
-#include "L1Trigger/L1TMuonBarrel/src/L1MuBMSecProcId.h"
+#include "DataFormats/L1TMuon/interface/L1MuBMSecProcId.h"
 
 //---------------
 // C++ Headers --
