+++ DataFormats/L1TMuon/src/classes.h
@@ -14,6 +14,13 @@
 #include "DataFormats/L1TMuon/interface/EMTFTrack2016Extra.h"
 #include "DataFormats/L1TMuon/interface/EMTFHit2016Extra.h"
 
+#include "DataFormats/L1TMuon/interface/L1MuBMTrack.h"
+#include "DataFormats/L1TMuon/interface/L1MuBMTrackSegPhi.h"
+#include "DataFormats/L1TMuon/interface/L1MuBMTrackSegEta.h"
+#include "DataFormats/L1TMuon/interface/L1MuBMSecProcId.h"
+#include "DataFormats/L1TMuon/interface/L1MuBMAddressArray.h"
+#include "DataFormats/L1TMuon/interface/L1MuBMTrackSegLoc.h"
+
 #include <vector>
 
 namespace {
@@ -23,32 +30,49 @@ namespace {
 
     l1t::RegionalMuonCandBxCollection regCand;
     edm::Wrapper<l1t::RegionalMuonCandBxCollection> regCandWrap;
-   
+
     l1t::EMTFDaqOutCollection emtfOutput;
     edm::Wrapper<l1t::EMTFDaqOutCollection> emtfOutputWrap;
-   
+
     l1t::EMTFHitCollection emtfHit;
     edm::Wrapper<l1t::EMTFHitCollection> emtfHitWrap;
-   
+
     l1t::EMTFRoadCollection emtfRoad;
     edm::Wrapper<l1t::EMTFRoadCollection> emtfRoadWrap;
-   
+
     l1t::EMTFTrackCollection emtfTrack;
     edm::Wrapper<l1t::EMTFTrackCollection> emtfTrackWrap;
-   
+
     l1t::EMTFTrack2016Collection emtfTrack2016;
     edm::Wrapper<l1t::EMTFTrack2016Collection> emtfTrack2016Wrap;
-   
+
     l1t::EMTFHit2016Collection emtfHit2016;
     edm::Wrapper<l1t::EMTFHit2016Collection> emtfHit2016Wrap;
-   
+
     l1t::EMTFTrack2016ExtraCollection emtfTrack2016Extra;
     edm::Wrapper<l1t::EMTFTrack2016ExtraCollection> emtfTrack2016ExtraWrap;
-   
+
     l1t::EMTFHit2016ExtraCollection emtfHit2016Extra;
     edm::Wrapper<l1t::EMTFHit2016ExtraCollection> emtfHit2016ExtraWrap;
-   
+
   };
 }
 
+namespace L1Trigger_L1TMuonBarrel {
+  struct dictionary {
+    L1MuBMTrackSegPhi l1mu_trk_ph;
+    L1MuBMTrackSegEta l1mu_trk_th;
+    L1MuBMTrack       l1mu_trk_tr;
+    L1MuBMSecProcId   l1mu_dt_proc;
+    L1MuBMTrackSegLoc  l1mu_dt_segloc;
+    L1MuBMAddressArray l1mu_dt_addr;
 
+    std::vector<L1MuBMTrackSegPhi> l1mu_trk_ph_V;
+    std::vector<L1MuBMTrackSegEta> l1mu_trk_th_V;
+    std::vector<L1MuBMTrack>       l1mu_trk_tr_V;
+
+    edm::Wrapper<std::vector<L1MuBMTrackSegPhi> > l1mu_trk_ph_W;
+    edm::Wrapper<std::vector<L1MuBMTrackSegEta> > l1mu_trk_th_W;
+    edm::Wrapper<std::vector<L1MuBMTrack> >       l1mu_trk_tr_W;
+  };
+}
