+++ L1Trigger/L1TMuonBarrel/src/L1MuBMTrackFinder.cc
@@ -41,12 +41,11 @@
 #include "L1Trigger/L1TMuonBarrel/src/L1MuBMEtaProcessor.h"
 #include "L1Trigger/L1TMuonBarrel/src/L1MuBMWedgeSorter.h"
 #include "L1Trigger/L1TMuonBarrel/src/L1MuBMMuonSorter.h"
-#include "L1Trigger/L1TMuonBarrel/interface/L1MuBMTrack.h"
 
 #include "DataFormats/L1TMuon/interface/RegionalMuonCand.h"
-#include "L1Trigger/L1TMuonBarrel/interface/L1MuBMTrack.h"
-#include "L1Trigger/L1TMuonBarrel/src/L1MuBMTrackSegPhi.h"
-#include "L1Trigger/L1TMuonBarrel/src/L1MuBMTrackSegEta.h"
+#include "DataFormats/L1TMuon/interface/L1MuBMTrack.h"
+#include "DataFormats/L1TMuon/interface/L1MuBMTrackSegPhi.h"
+#include "DataFormats/L1TMuon/interface/L1MuBMTrackSegEta.h"
 
 using namespace std;
